Create exectuor interface and LocalExecutor for interchangable executors
