+++ pkg/runtime/executor.go
@@ -1,0 +1,5 @@
+package runtime
+
+type Executor interface {
+	Execute(test TestCase) TestResult
+}
