+++ pkg/runtime/runtime_test.go
@@ -1,9 +1,7 @@
 package runtime
 
 import (
-	"fmt"
 	"github.com/stretchr/testify/assert"
-	"runtime"
 	"testing"
 	"time"
 )
@@ -61,61 +59,6 @@ func TestRuntime_WithRetriesAndInterval(t *testing.T) {
 
 	assert.Equal(t, 1, counter)
 	assert.True(t, duration.Seconds() > 0.15, "Retry interval did not work")
-}
-
-func TestRuntime_WithEnvVariables(t *testing.T) {
-	envVar := "$KEY"
-	if runtime.GOOS == "windows" {
-		envVar = "%KEY%"
-	}
-
-	s := TestCase{
-		Command: CommandUnderTest{
-			Cmd:     fmt.Sprintf("echo %s", envVar),
-			Timeout: "2s",
-			Env:     map[string]string{"KEY": "value"},
-		},
-		Expected: Expected{
-			Stdout: ExpectedOut{
-				Contains: []string{"value"},
-			},
-			ExitCode: 0,
-		},
-		Title: "Output env variable",
-	}
-
-	got := runTest(s)
-	assert.True(t, got.ValidationResult.Success)
-}
-
-func Test_runTestShouldReturnError(t *testing.T) {
-	test := TestCase{
-		Command: CommandUnderTest{
-			Cmd: "pwd",
-			Dir: "/home/invalid",
-		},
-	}
-
-	got := runTest(test)
-
-	if runtime.GOOS == "windows" {
-		assert.Contains(t, got.TestCase.Result.Error.Error(), "chdir /home/invalid")
-	} else {
-		assert.Equal(t, "chdir /home/invalid: no such file or directory", got.TestCase.Result.Error.Error())
-	}
-}
-
-func TestRuntime_WithInvalidDuration(t *testing.T) {
-	test := TestCase{
-		Command: CommandUnderTest{
-			Cmd:     "echo test",
-			Timeout: "600lightyears",
-		},
-	}
-
-	got := runTest(test)
-
-	assert.Equal(t, "time: unknown unit lightyears in duration 600lightyears", got.TestCase.Result.Error.Error())
 }
 
 func getExampleTestSuite() []TestCase {
