+++ pkg/runtime/ssh_executor.go
@@ -1,0 +1,8 @@
+package runtime
+
+type SSHExecutor struct {
+	Host       string
+	User       string
+	Password   string
+	PrivateKey string
+}
